更新2023.8.23(暑假最后一次) - 修复一些bug - + HiddenTopWindow移动到别的element时会闪一下 - + 色块（IntUI#colorBlock）的alphaBg的透明度不恒为1（随color的alpha） - + 修复Tester排序的bug和多tester时context为null的bug - 重构一些代码，优化代码 - + UnitSpawn添加menu快速调用方法（原先只能在设置里点击执行） - + Tester历史和收藏夹添加搜索功能 - + 重构ValueLabel - 修改stopIfOvertime的icon - JSSyntax添加变量高亮 - 添加Executor，可以添加在收藏夹长按，添加启动js - JSFunc（js中IntFunc）添加更多方法 - 添加Interp显示，ShowUIList支持颜色保存 - 一些细节上的优化

+++ src/modtools/ModTools.java
@@ -2,25 +2,32 @@ package modtools;
 
 import arc.*;
 import arc.files.*;
+import arc.graphics.Color;
 import arc.struct.*;
 import arc.util.*;
 import arc.util.io.PropertiesUtils;
 import dalvik.system.BaseDexClassLoader;
 import ihope_lib.MyReflect;
-import mindustry.Vars;
+import mindustry.*;
 import mindustry.game.EventType.ClientLoadEvent;
+import mindustry.graphics.LoadRenderer;
 import mindustry.mod.*;
-import mindustry.mod.Mods.LoadedMod;
+import mindustry.mod.Mods.*;
 import modtools.graphics.MyShaders;
 import modtools.ui.*;
+import modtools.ui.content.debug.Tester;
 import modtools.ui.tutorial.AllTutorial;
 import modtools.utils.Tools;
+import modtools.utils.reflect.FieldUtils;
+import modtools.utils.ui.FileUtils;
 
-import java.io.File;
+import java.io.*;
+import java.lang.reflect.Field;
 import java.net.URLClassLoader;
 import java.util.Arrays;
 
 import static mindustry.Vars.ui;
+import static modtools.IntVars.root;
 import static modtools.utils.MySettings.SETTINGS;
 
 public class ModTools extends Mod {
@@ -33,32 +40,35 @@ public class ModTools extends Mod {
 		}
 	}
 
-	@SuppressWarnings("Convert2MethodRef")
 	public ModTools() {
 		Log.info("Loaded ModTools constructor" + (isImportFromGame ? " [[[from game]]]" : "") + ".");
 
-		if (isImportFromGame) try {
-			resolveLibs();
-			resolveUI();
-		} catch (Throwable e) {
-			Vars.ui.showException("Cannot load ModTools. (Don't worry)", e);
-		}/* !isImportFromGame */
-		else {
-			Tools.forceRun(() -> {
-				if (Vars.mods.getMod(ModTools.class) == null) return false;
-				Log.info("ok");
-				resolveLibs();
-				return true;
-			});
-			Events.on(ClientLoadEvent.class, e -> resolveUI());
-		}
+		ObjectMap<Class<?>, ModMeta> metas = Reflect.get(Mods.class, Vars.mods, "metas");
+		IntVars.meta = metas.get(ModTools.class);
 
+		try {
+			if (!isImportFromGame) IntVars.meta.hidden = false;
+			resolveLibsCatch();
+
+			if (isImportFromGame) resolveUI();
+			else Events.on(ClientLoadEvent.class, e -> resolveUI());
+		} catch (Throwable e) {
+			if (isImportFromGame) Vars.ui.showException("Cannot load ModTools. (Don't worry.)", e);
+			Log.err("Failed to load ModTools.", e);
+		}
+	}
+
+
+	public void loadContent() {
+		IntVars.meta.hidden = true;
+		Core.app.post(Tester::initExecution);
 	}
 	public static Fi findRoot() {
 		if (OS.isWindows) {
 			return Fi.get(((URLClassLoader) ModTools.class.getClassLoader()).getURLs()[0].getFile());
 		}
-		return findRootAndroid();
+		if (OS.isAndroid) return findRootAndroid();
+		throw new UnsupportedOperationException();
 	}
 	public static Fi findRootAndroid() {
 		Object pathList = Reflect.get(BaseDexClassLoader.class, ModTools.class.getClassLoader(),
@@ -67,25 +77,80 @@ public class ModTools extends Mod {
 		File     file = Reflect.get(arr[0], "path");
 		return new Fi(file);
 	}
+
+	private static void resolveLibsCatch() {
+		try {
+			resolveLibs();
+		} catch (UnexpectedPlatform e) {
+			Log.err("It seems you platform is special. (But don't worry.)");
+			Tools.forceRun(() -> {
+				if (Vars.mods.getMod(ModTools.class) == null) return false;
+				resolveLibs();
+				return true;
+			});
+		}
+	}
 	private static void resolveLibs() {
 		LoadedMod mod = Vars.mods.getMod(ModTools.class);
 		root = mod != null ? mod.root : new ZipFi(findRoot());
-		mainLoader = (ModClassLoader) Vars.mods.mainLoader();
 		libs = root.child("libs");
 
 		loadLib("reflect-core", "ihope_lib.MyReflect", true, () -> MyReflect.load());
+		setLoadRenderer();
 		IntVars.hasDecompiler = loadLib("procyon-0.6", "com.strobel.decompiler.Decompiler", false);
 		if (isImportFromGame) loadBundle();
 
 		// MyReflect.classDefiner().defineClass();
-		MyShaders.load();
 	}
+
+	private static void setLoadRenderer() {
+		try {
+			Time.mark();
+			Field        field  = FieldUtils.getFieldAccess(ClientLauncher.class, "loader");
+			LoadRenderer loader = (LoadRenderer) field.get(Vars.platform);
+			Color        color  = Reflect.get(LoadRenderer.class, loader, "color");
+			color.set(Color.cyan);
+			color = Reflect.get(LoadRenderer.class, loader, "colorRed");
+			color.set(Color.sky);
+			FieldUtils.setValue(LoadRenderer.class.getDeclaredField("orange"),
+			 loader, "[cyan]");
+			FieldUtils.setValue(LoadRenderer.class.getDeclaredField("red"),
+			 loader, "[sky]");
+
+			// FieldUtils.setValue(Enum.class.getDeclaredField("ordinal"), LogLevel.warn, 0);
+		} catch (Exception e) {
+			Log.err(e);
+		} finally {
+			Log.info("Time to set load renderer: @ms", Time.elapsed());
+		}
+	}
+
 	private static void resolveUI() {
-		if (throwable != null && ui != null) {
-			ui.showException(throwable);
+		if (ui == null) return;
+		if (error != null) {
+			ui.showException(error);
 			return;
 		}
-
+		if (OS.isWindows) ui.mods.shown(() -> {
+			ui.mods.buttons.row().button("importFromSelector", () -> {
+				FileUtils.openFiSelector(list -> {
+					try {
+						for (Fi fi : list) {
+							if (!fi.extEquals("zip") && !fi.extEquals("jar"))
+								throw new IllegalArgumentException("Unexpected file type: " + fi.extension());
+							try {
+								Vars.mods.importMod(fi);
+							} catch (IOException e) {
+								throw new RuntimeException(e);
+							}
+						}
+					} catch (Throwable e) {
+						IntUI.showException("Failed to import mod from selector", e);
+					}
+				});
+			});
+		});
+		MyShaders.load();
 		Time.runTask(6f, () -> {
 			IntUI.load();
 			AllTutorial.init();
@@ -94,6 +159,7 @@ public class ModTools extends Mod {
 			Tools.runIgnoredException(Background::main);
 		}
 	}
+
 	/**
 	 * @see Mods#buildFiles()
 	 */
@@ -128,10 +194,8 @@ public class ModTools extends Mod {
 
 	}
 
-	public static ModClassLoader mainLoader;
-	public static Fi             root;
-	public static Throwable      throwable = null;
-	public static Fi             libs;
+	public static Throwable error = null;
+	public static Fi        libs;
 
 	public static boolean loadLib(String fileName, String mainClassName, boolean showError) {
 		return loadLib(fileName, mainClassName, showError, null);
@@ -140,54 +204,37 @@ public class ModTools extends Mod {
 	public static boolean loadLib(String fileName, String mainClassName, boolean showError, Runnable callback) {
 		try {
 			// 没报错的话，证明已经加载
-			mainLoader.loadClass(mainClassName);
+			IntVars.mainLoader.loadClass(mainClassName);
 			return true;
 		} catch (Exception ignored) {}
 
+		Fi sourceFi = libs.child(fileName + ".jar");
+		if (!sourceFi.exists()) return false;
+
 		Log.info("Loading @.jar", fileName);
+		Time.mark();
 		// 加载反射
 		try {
-			Fi sourceFi = libs.child(fileName + ".jar");
-			Log.info("Loading source fi: " + sourceFi);
 			Fi toFi = Vars.dataDirectory.child("tmp/mod-tools-" + fileName + ".jar");
-			delete(toFi);
+			IntVars.delete(toFi);
 			sourceFi.copyTo(toFi);
-			ClassLoader loader = Vars.platform.loadJar(toFi, mainLoader);
-			mainLoader.addChild(loader);
+			ClassLoader loader = Vars.platform.loadJar(toFi, IntVars.mainLoader);
+			IntVars.mainLoader.addChild(loader);
 			Class.forName(mainClassName, true, loader);
 			if (callback != null) callback.run();
 			toFi.delete();
 			return true;
 		} catch (Throwable e) {
 			if (showError) {
-				throwable = e;
-				Log.err(e);
+				error = e;
+				Log.err("Unexpected exception when loading '" + sourceFi + "'", e);
 			}
 			return false;
+		} finally {
+			Log.info("Loaded '@' in @ms", sourceFi, Time.elapsed());
 		}
 	}
-	/** @return {@code true} if the file be deleted successfully */
-	private static boolean delete(Fi fi) {
-		return fi.exists() && (fi.isDirectory() ? fi.deleteDirectory() : fi.delete());
-	}
-	/*public static void test() {
-		new Thread(() -> Core.app.exit());
-		unsafe.park(false, (long) 1E9);
-		try {
-			Method m = Application.class.getDeclaredMethod("getListeners");
-			m.setAccessible(true);
-			Seq<ApplicationListener> listeners = (Seq<ApplicationListener>) m.invoke(Core.app);
-			listeners.remove((ApplicationListener) Vars.platform);
-			Field field = SdlApplication.class.getDeclaredField("running");
-			field.setAccessible(true);
-			m = SdlApplication.class.getDeclaredMethod("loop");
-			m.setAccessible(true);
-			while (field.getBoolean(Core.app)) {
-				m.invoke(Core.app);
-			}
-			Log.info("finish");
-		} catch (Throwable e) {
-			Log.err(e);
-		}
-	}*/
+
+	static class UnexpectedPlatform extends RuntimeException {}
+
 }
